# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/flights/amadeus/search": {
      "get": {
        "operationId": "FlightsController_amadeusSearch",
        "parameters": [
          {
            "name": "originLocationCode",
            "required": true,
            "in": "query",
            "description": "Origin IATA airport/city code",
            "schema": {
              "example": "SYD",
              "type": "string"
            }
          },
          {
            "name": "destinationLocationCode",
            "required": true,
            "in": "query",
            "description": "Destination IATA airport/city code",
            "schema": {
              "example": "BKK",
              "type": "string"
            }
          },
          {
            "name": "departureDate",
            "required": true,
            "in": "query",
            "description": "Departure date (YYYY-MM-DD)",
            "schema": {
              "example": "2025-08-01",
              "type": "string"
            }
          },
          {
            "name": "returnDate",
            "required": false,
            "in": "query",
            "description": "Return date for round-trip (YYYY-MM-DD)",
            "schema": {
              "example": "2025-08-15",
              "type": "string"
            }
          },
          {
            "name": "adults",
            "required": true,
            "in": "query",
            "description": "Number of adult travelers",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "children",
            "required": false,
            "in": "query",
            "description": "Number of child travelers",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "infants",
            "required": false,
            "in": "query",
            "description": "Number of infant travelers",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "travelClass",
            "required": false,
            "in": "query",
            "description": "Travel class",
            "schema": {
              "enum": [
                "ECONOMY",
                "PREMIUM_ECONOMY",
                "BUSINESS",
                "FIRST"
              ],
              "type": "string"
            }
          },
          {
            "name": "includedAirlineCodes",
            "required": false,
            "in": "query",
            "description": "Comma-separated airline IATA codes to include",
            "schema": {
              "example": "TG",
              "type": "string"
            }
          },
          {
            "name": "excludedAirlineCodes",
            "required": false,
            "in": "query",
            "description": "Comma-separated airline IATA codes to exclude",
            "schema": {
              "example": "AA,TP",
              "type": "string"
            }
          },
          {
            "name": "nonStop",
            "required": false,
            "in": "query",
            "description": "Only non-stop flights",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "currencyCode",
            "required": false,
            "in": "query",
            "description": "Currency code for pricing",
            "schema": {
              "example": "USD",
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "example": 1500,
              "type": "number"
            }
          },
          {
            "name": "max",
            "required": false,
            "in": "query",
            "description": "Max number of offers to return",
            "schema": {
              "default": 10,
              "example": 5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search flights via Amadeus (GET)",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      },
      "post": {
        "operationId": "FlightsController_amadeusSearchPost",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Search flights via Amadeus (POST — advanced)",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/price": {
      "post": {
        "operationId": "FlightsController_amadeusPrice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmadeusPriceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Price a flight offer",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/order": {
      "post": {
        "operationId": "FlightsController_amadeusOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmadeusOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a flight order (booking)",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/order/{orderId}": {
      "get": {
        "operationId": "FlightsController_amadeusGetOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a flight order by ID",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      },
      "delete": {
        "operationId": "FlightsController_amadeusDeleteOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel/delete a flight order",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/seatmap": {
      "get": {
        "operationId": "FlightsController_amadeusSeatmapByOrder",
        "parameters": [
          {
            "name": "flightOrderId",
            "required": true,
            "in": "query",
            "description": "Amadeus flight order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get seatmap by flight order ID",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      },
      "post": {
        "operationId": "FlightsController_amadeusSeatmapByOffer",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Get seatmap by flight offer (POST)",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
...
    "/api/v1/flights/amadeus/price-analysis": {
      "get": {
        "operationId": "FlightsController_amadeusPriceAnalysis",
        "parameters": [
          {
            "name": "originIataCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MAD",
              "type": "string"
            }
          },
          {
            "name": "destinationIataCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "CDG",
              "type": "string"
            }
          },
          {
            "name": "departureDate",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-08-01",
              "type": "string"
            }
          },
          {
            "name": "currencyCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "EUR",
              "type": "string"
            }
          },
          {
            "name": "oneWay",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight price analysis (itinerary price metrics)",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
...
    "/api/v1/flights/amadeus/inspiration": {
      "get": {
        "operationId": "FlightsController_amadeusInspiration",
        "parameters": [
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "duration",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "oneWay",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "departureDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "origin",
            "required": true,
            "in": "query",
            "schema": {
              "example": "BOS"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight inspiration search — cheapest destinations",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/cheapest-dates": {
      "get": {
        "operationId": "FlightsController_amadeusCheapestDates",
        "parameters": [
          {
            "name": "departureDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "destination",
            "required": true,
            "in": "query",
            "schema": {
              "example": "LON"
            }
          },
          {
            "name": "origin",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MAD"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight cheapest date search",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
...
    "/api/v1/flights/amadeus/recommendations": {
      "get": {
        "operationId": "FlightsController_amadeusRecommendations",
        "parameters": [
          {
            "name": "travelerCountryCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "FR"
            }
          },
          {
            "name": "cityCodes",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PAR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Travel recommendations",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/status": {
      "get": {
        "operationId": "FlightsController_amadeusFlightStatus",
        "parameters": [
          {
            "name": "scheduledDepartureDate",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-08-01"
            }
          },
          {
            "name": "flightNumber",
            "required": true,
            "in": "query",
            "schema": {
              "example": "986"
            }
          },
          {
            "name": "carrierCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "BA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "On demand flight status",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
...
    "/api/v1/flights/amadeus/airport-ontime": {
      "get": {
        "operationId": "FlightsController_amadeusAirportOnTime",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-08-01"
            }
          },
          {
            "name": "airportCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "JFK"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airport on-time performance",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airports": {
      "get": {
        "operationId": "FlightsController_amadeusAirportSearch",
        "parameters": [
          {
            "name": "countryCode",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MUC"
            }
          },
          {
            "name": "subType",
            "required": true,
            "in": "query",
            "schema": {
              "example": "CITY,AIRPORT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airport & city search by keyword",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airports/nearest": {
      "get": {
        "operationId": "FlightsController_amadeusNearestAirports",
        "parameters": [
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2.55,
              "type": "number"
            }
          },
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 49,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Nearest relevant airports",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airports/routes": {
      "get": {
        "operationId": "FlightsController_amadeusAirportRoutes",
        "parameters": [
          {
            "name": "max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departureAirportCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MAD"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airport direct destinations / routes",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airports/{locationId}": {
      "get": {
        "operationId": "FlightsController_amadeusAirportById",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airport & city search by ID",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airlines/checkin-links": {
      "get": {
        "operationId": "FlightsController_amadeusCheckinLinks",
        "parameters": [
          {
            "name": "airlineCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "IB"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight check-in links",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airlines/lookup": {
      "get": {
        "operationId": "FlightsController_amadeusAirlineLookup",
        "parameters": [
          {
            "name": "airlineCodes",
            "required": true,
            "in": "query",
            "schema": {
              "example": "BA,AIC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airline code lookup",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/amadeus/airlines/routes": {
      "get": {
        "operationId": "FlightsController_amadeusAirlineRoutes",
        "parameters": [
          {
            "name": "max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "airlineCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "AF"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Airline routes / destinations",
        "tags": [
          "Flights",
          "Amadeus — Flight Booking"
        ]
      }
    },
    "/api/v1/flights/search": {
      "post": {
        "operationId": "FlightsController_search",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchFlightsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Search flights from local database",
        "tags": [
          "Flights"
        ]
      }
    },
    "/api/v1/flights/popular": {
      "get": {
        "operationId": "FlightsController_getPopular",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get popular/featured flights",
        "tags": [
          "Flights"
        ]
      }
    },
    "/api/v1/flights/deals": {
      "get": {
        "operationId": "FlightsController_getDeals",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get flight deals (cheapest)",
        "tags": [
          "Flights"
        ]
      }
    },
...
  },
  "info": {
    "title": "✈️ Flight Booking API",
    "description": "\n      ## Multi-Tenant Flight Booking Platform API\n      \n      A scalable flight booking backend with:\n      - **Multi-tenancy** (B2B & B2C)\n      - **Multi-currency** support (14 currencies)\n      - **Dual payment** processing (Stripe + Paystack)\n      - **Real-time** seat availability\n      - **Admin analytics** and reporting\n      - **Global Travel Integration** (Amadeus Search, Price, Booking)\n      \n      ### Amadeus API Modules\n      This platform integrates the following Amadeus modules:\n      - **Flights**: Live search, pricing, and orders (31 endpoints)\n      - **Hotels**: Search, offers, and bookings (10 endpoints)\n      - **Experiences**: Tours, activities, and city search (4 endpoints)\n      - **Transfers**: Cars and transfers search/booking (3 endpoints)\n      - **Insights**: Market trends and traffic analytics (3 endpoints)\n      - **Itinerary**: Travel predictions and planning (1 endpoint)\n      \n      ### Authentication\n      All protected endpoints require a Bearer JWT token.\n      Use `/auth/login` to obtain tokens.\n      \n      ### Rate Limiting\n      API requests are rate-limited to 100 requests per minute per IP.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}